testing riot api on heruku
